requirements.txt
Flask
Flask-PyMongo
Flask-JWT-Extended
Flask-Mail
python-dotenv
marshmallow
pymongo
